streamlit_app/requirements.txt
streamlit==1.22.0
pandas==2.0.3
matplotlib==3.7.2
pillow==9.5.0
opencv-python==4.8.0.74
pytesseract==0.3.10
transformers==4.31.0
